+++ src/header/index.js
@@ -37,21 +37,21 @@ const Headermain = () => {
                   <li className="menu_item ">
                     <Link  onClick={handleToggle} to="/" className="my-3">Home</Link>
                   </li>
+                  {/* <li className="menu_item">
+                    <Link onClick={handleToggle} to="/about" className="my-3">About</Link>
+                  </li> */}
                   <li className="menu_item">
                     <Link  onClick={handleToggle} to="/portfolio" className="my-3"> Design</Link>
-                  </li>
-                  <li className="menu_item">
-                    <Link  onClick={handleToggle} to="/news" className="my-3"> News</Link>
-                  </li>
-                  <li className="menu_item">
-                    <Link onClick={handleToggle} to="/about" className="my-3">About</Link>
                   </li>
                   <li className="menu_item">
                     <Link onClick={handleToggle} to="/research" className="my-3"> Research</Link>
                   </li>
                   <li className="menu_item">
-                    <Link onClick={handleToggle} to="/contact" className="my-3"> Contact</Link>
+                    <Link  onClick={handleToggle} to="/news" className="my-3"> News</Link>
                   </li>
+                  {/* <li className="menu_item">
+                    <Link onClick={handleToggle} to="/contact" className="my-3"> Contact</Link>
+                  </li> */}
 
                 </ul>
               </div>