+++ src/pages/home/index.js
@@ -71,8 +71,8 @@ export const Home = () => {
           <div className="text order-2 order-lg-1 h-100 d-lg-flex justify-content-center">
             <div className="align-self-center ">
               <div className="intro mx-auto">
-                <h2 className="mb-1x">{introdata.title}</h2>
-                <h1 className="fluidz-48 mb-1x">
+                <h1 className="mb-1x">{introdata.title}</h1>
+                <h2 className="fluidz-48 mb-1x">
                   <Typewriter
                     options={{
                       strings: [
@@ -85,7 +85,7 @@ export const Home = () => {
                       deleteSpeed: 10,
                     }}
                   />
-                </h1>
+                </h2>
                 <p className="mb-1x">{introdata.description}</p>
                 <div className="intro_btn-action pb-5">
                   <Link to="/portfolio" className="text_2">
@@ -96,9 +96,9 @@ export const Home = () => {
                       <div className="ring three"></div>
                     </div>
                   </Link>
-                  <Link to="/contact">
+                  <Link to="/research">
                     <div id="button_h" className="ac_btn btn">
-                      Contact Me
+                      My Research
                       <div className="ring one"></div>
                       <div className="ring two"></div>
                       <div className="ring three"></div>